+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+
 # 0.0.1 
 * Initial Release
 
@@ -11,3 +12,8 @@
 # 0.0.4 
 
 * Supports returning only a single value, such as "host", "instance_id", ect
+
+# 0.0.5
+
+* Much more meta data from running instance
+* Pretty print mode